Fixes #2
The issue was that I (not Piet) reimplemented Brückmann's formula incorrectly in characteristic 0. The formula in 5.1 of Satz 5 of the second complete intersections paper has 3 terms, and the last two are omitted in characteristic 0. I mistakenly only omitted the second term. I was not born to be a Lisp programmer is the real upshot here.